templates/ml-pipeline/requirements.txt
numpy==1.26.4
pandas==2.2.0
scikit-learn==1.4.0
mlflow==2.12.0
jupyter==1.0.0
jupyterlab==4.1.0
matplotlib==3.8.0
seaborn==0.13.0
pyyaml==6.0.1
pytest==8.2.0
